[Runner] Fix a concurrency issue
The underlying cause is the listener's onError method setting the
state to DISCONNECTED, calling the status listener which called
updateDispatching. This was bad, as the concurrent binary output
stream writing also happened in the synchronized method
updateDispatching, causing a deadlock. The listener never exited,
never releasing the resources and therefore the write couldn't exit
either?