Fix CommitInfoTable formatting and highlighting colour
The table previously displayed "NaN" when there was no change (i.e. the
previous commit did not share the measurement or it failed there). This
was because it could not parse "-" as a number.
Now it gets passed the raw number and "-" is filtered out beforehand.

A second problem was that the highlighting was off. Sadly the method got
passed the *formatted* number, which could no longer be parsed
correctly. The comparisons therefore compared with NaN, which was not
really helpful.
Now the method gets the raw, unformatted number and does the formatting
itself.