[Backend] GuickCloning: Clone submodules correctly with JGit
This probably wasted a few hours because JGit seems to be god damn
stupid. My submodule name was not equal to my submodule path, causing
the JGit lookups to fail. Native git had no problems with that though
as there is a name switch in the "submodule add" command...

I guess I will never know why nobody implemented that and the only
documentation was a *link to git's man page*.

So thanks, I guess. Why did we even use JGit to start with? I am not
sure my desk is up to the task of fixing the last JGit bugs...

Next project is *definitely* switching fetch to native git when
available.