[Runner] Maybe fix runner deadlock
The runner might have gotten two aborts right after each other. If this
happens, the second abort will see that the runner is not busy and instantly
complete, *setting the cancelled flag in the process*. Due to this the next
startWork command will silently exit immediately and not send any results,
leaving the runner stranded in an executing state.