[backend] Nest run inserts into a transaction
In rare circumstances where velcom crashes right after it inserted a run into the database, but before it inserted any of the measurements and measurement values, an inconsistent state in the database was possible.

To prevent this, runs are now inserted along with their respective measurements and measurement values using a transaction.