Adapt frontend to the domain at KIT.
Unfortunately, this requires to separate Docker images since the domain of
the backend is baked in when building the frontend. The convention is to
leave the packages for aaaaaaah.de untouched and add "-kit" to the
packges for kit.edu.

TODO: There is a lot of duplication between the jobs "push" and "push-kit".