[frontend] Fix public path settings
The router uses history mode and therefore needs to painfully aware of
where the SPA is served from. In a previous commit support for
deployment at a subpath was introduced, but the BASE_URL parameter in
the router was missed.

This commit removes the "SITE" parameter introduced earlier and replaces
it with BASE_URL. This ensures the router is kept in sync with the
publicPath.