[backend] Introduce structures for kicking disconnected runners
Runners will be kicked if they are disconnected for too long, so they do
not appear in the UI anymore.

This patch adds the necessary mechanics to make that work, a later patch
will introduce a watchdog actually calling the cleanup methods.
Disposing runners safely after their grace period is up is a bit hairy,
as a new runner could join at any time and potentially use the now
disposed runner. This would lead to a "Zombie-Runner" that is not known
to the dispatcher but can still receive work.