[backend] Re-Send runner its current work when it asks again
If a runner asks for work while the backend has a pretty good idea of what
it should be doing, it will resend that work instead of asking the dispatcher
for the next task. Otherwise the task would get lost.

This should only happen when a runner restarts after it was assigned some
work and forgot it should be doing something. In every other scenario
this is illegal behaviour, but the backend tries to smooth it over here.