[frontend] Rework persistence
I will try to take the slightly nicer road: Storing objects in the
stores and (de)serializing them to JSON. This makes a few things more
complicated and requires some plumbing code, but the stores should be
easier to write and maintain.

This patch just starts on that work, the stores need to be adjusted and
some private variables can likely be removed.

This patch also grossly circumvents `private` while harming type safety,
but there is not yet a way around that. Eliminating the need for private
variables could help here.