[backend] Fix command timeout while transferring large repos
Before this patch the runner sent work synchronously to the websocket
thread, maybe preventing it from responding to pings.

Additionally, the periodic status requester tried to ask the runner for
its status *while the task/bench repo was send over*, causing a timeout
and kicking the runner.
The new state prevents that.

The last change was in the dispatcher registration, causing runners to
be added without having any information.