[backend] Fix cache meter name collision
Apparently (for whatever reason) prometheus or micrometer can't handle meters with the same name but differing tags. With this change names of meters that track caches will be modified, so that the name of the cache is prepended to the name of the meter so that this this collision problem no longer occurs.