[frontend] Add an option to echarts graph to show day-equidistant points
Using this commit's will be grouped into buckets using the day they were
authored on. Inside that bucket commits will be position equidistantly.
This means the x axis is a bit inaccurate, but it is a lot easier to see
exact relationships and the tooltips still display the correct value.