[backend] Always find all commits in Listener#checkForUnknownCommits()
Before this change, only commits reachable from tracked branches were marked as known (if the repo already has some known commits). Commits from untracked branches were completely ignored.

This caused an exception when a untracked branch is changed to tracked and a commit of this branch is inserted into the queue before the Listener had the chance to mark this commit as known.

This commit changes the Listener behaviour in such a way that now all commits (regardless if from a tracked or untracked branch) are at least marked as known. Only commits from tracked branches are then also inserted into the queue.