[frontend] Navigate to detail graph on click but ignore selection
This patch introduces a mechanism that allows clicking on rows in the
measurement table. If you do that, it will take you to the detail graph
with the commit and dimension you clicked on marked as the reference
commit.

This would normally interfere with text selection (e.g. copying the
whole table or just a single row). In order to not annoy the user,
letting go of the mouse will only navigate when document.getSelection()
is empty - i.e. the user didn't select any text! If they did select
something, the navigation will be cancelled and user shouldn't notice
the feature exists.