[backend] Cache available dimensions
Previously, the /all-repos endpoint took a few hundred ms, which is too much for
an endpoint that gets queried constantly. With this cache, the time is down to
about 10-15 ms when testing locally, which is more appropriate.

The cache introduced in this commit is also hopefully implemented in a way that
makes cache usage and invalidation across multiple access classes easier to
implement and understand without any access class depending on another access
class.