[frontend] Mitigate XSS vulnerability in graph tooltips
Vue does quite a lot to prevent (unwanted) XSS vulnerabilities and
mostly sets the inner*Text* and not the innerHTML (unless you use v-html
and force it to be dangerous).

The two graph libraries used allow you to specify a tooltip/legend
formatter producing raw HTML - which will be inserted as HTML and
completely sidesteps vue.
This works, but it also allows users to specify authors, dimensions or
commit summaries that can inject *arbitrary HTML* into the page.

On a whim I tried to exploit this with
https://github.com/IPDSnelting/velcom-test-repo/commit/6b9d662218eebfa7ba7a94433971e18dd847c68d
and was trivially able to pop up an alert dialog. Not good.

This PR fixes the graph tooltip handlers by only displaying *escaped*
user-submitted text.