[backend] Switch Telerunner/Dispatcher over to abort-polling
Before this patch the dispatcher was notified when a task was removed
from the queue. As this callback mechanism was removed in earlier
commits, we needed a new strategy.

Now the TeleRunner will, whenever it receives a status ping from the
runner, query the queue and ask whether the current task was aborted. If
so, it will trigger an abort on the runner.

This mechanism is simpler, can't miss updates and shouldn't be too
inefficient.