[backend] Sleep for 5 seconds between "get_status" requests
Before this patch the backend didn't sleep *at all* if the runner had no
result and if it had one it slept for a very, very short duration due to
a unit mismatch. This leads to a lot of useless log output and CPU usage
in the runner and backend.