[backend] Try to make the cleansUpDisconnectedRunner test less flaky
Previously if the stars aligned just right, the background cleaner task
could remove the runner after it was added but *before* the next line in
the test, checking that the runner was added, was run.

This was possible as the runner *always* return an invalid ping time,
even before it was added. This patch adjusts it to only return an
invalid time once we verified the add was successful.