[runner] Attempt to fix fd leak
A bug report showed the runner leaking file descriptors when doing lots of
reconnection attempts.

Apparently, HttpClient only cleans itself up if there are no more references to
it. In practice, this means that it only cleans itself up once a few other
objects with references to it have been garbage collected. However, it can take
quite a while for this to happen "naturally".

For each HttpClient or associated WelSocket (I'm not sure which) a few file
descriptors are opened. These are not freed immediately after a connection
attempt fails, but instead only after GC (as described above). When testing with
a reconnection delay of 0.1 seconds, velcom at one point managed to hold over
9,000 file descriptors.

This commit "fixes" the issue by kicking off GC manually via System.gc()
whenever a connection is closed. Another alternative might've been to call
the (private) cleanup methods via reflection, but that would've been even
uglier.