[frontend] Don't show results for unknown dimensions
If you:
1. Selected a repo
2. Selected a dimension
3. Selected another repo *without* that dimension
4. Deselected the original one

the frontend would request results for the originally selected
dimension, even if the second repo doesn't share it. The backend
responded with a bunch of "unbenchmarked metric" results, which were
then shown in the graph.

This patch fixes that by not requesting data for dimensions the repo
doesn't share.