[frontend] Hide series without values in Dygraphs graph
The repo comparison returns a bunch of commits with exactly one value.
This conflicts with the structure of the graphs before this change:
They expect every data point to have values for *every* possible series.
This lead to duplicated lines in the graph, as every datapoint would
show points for *all* repos, not just the one it belongs to.
Additionally the points might get weird values, as the graph tried to
find a suitable value for them.

This patch lets the dygraphs graph skip points for a series, if it
doesn't have *any* value for it.