[frontend] Navigate to base repo-detail after deleting a repo
The delete logic still used the old route for the repo detail page,
leading to a quite weird navigation state that wasn't quite the repo
detail page but also not a 404. This patch avoids that by updating the
repo detail route name which now properly shows the repo detail page
after the deletion.