[frontend] Add a share button to the ComparisonGraph
The share button does generate a valid link, but its terminology isn't
really correct right now, as it is the same one used for the detail
graph. Additionally, no "adjust to url" method is implemented yet.

This commit tries to consolidate both share dialogs into a single
component. If this proves too difficult the underlying mechanism
apparently was too specialized and we will need to split it up in two
components.