[frontend] Properly round link dates
VelCom rounds graph boundaries to the nearest midnight: the start time
is always rounded down and the end time up.
This ensures that every commit is visible, even if a day-equidistant
graph is used.

Before this patch the rounding logic was quite weird and didn't really
do what you expected. This patch simplifies it and fixes a few corner
cases (e.g. roundUp didn't properly round up as 23:00 was left as
23:00). Additionally the rounding functions now have fixpoints at
midnight, so repeated calls do not change the date. This is nice as a
date might be normalized more than once and that no longer changes the
date.