[frontend] Add a full-screen mode to dimension selection
A previous patch moved the dimension selection to the left of the graph.
This could lead to some unfortunate UX degradations, as the matrix
selector absolutely needs quite a lot of horizontal space. To combat
this, the patch introduces a full-screen mode, overlaying the selector
over the graph. After this change you are able to make your changes in
full-screen while still seeing the effects through the semi-transparent
background.