[frontend] Change to a slightly more sane persistence solution
Previously persistence was handled in the StorePersistenceUtilities
which casted things to any, tried to hydrate individual points and
classes and just generated a whole lot of mildly cooked spaghetti.

This patch hand-rolls a small persistence solution using the replacer
and restorer functions you can pass to JSON.stringify and JSON.parse.
These functions just take care to include the name of constructors when
serialising in the replacer function and then look up known constructor
names when deserialising in the restorer function.

With this solution JSON.parse/stringify will take care of recursively
restoring objects in a useful order, simplifying the code and making it
more robust to changes.

Additionally, the stores now have a "toPlainObject" function which
converts them to a normal object that should be serialized. With this
change, the stores filter their properties and can access their own
private properties without any `any` casting! This allows us to rename
properties or remove them and get nice little compiler errors instead of
random runtime crashes.