[frontend] Use checkboxes in graph setting labels
Before this graph the test of the button changed (i.e. "Begin Y-Axis at
minimum value" -> "Begin Y-Axis at zero). This caused the width of the
button to vary considerably, causing page reflows and making the UI a
bit more confusing.

Now the wording stays the same and a checkbox icon conveys the
information and the width of each button stays consistent.