[frontend] Group graph settings buttons on all screen sizes
Before this patch each button was its own column and a spacer provided
the necessary gap between the graph type (fast / fancy) and the buttons.
This caused interesting reflow behaviours, as parts of the buttons would
get moved down and out of alignment.

This patch creates two cols and sets the row to justify=space-between,
to keep this more consistent.