[frontend] Normalize to different repo when baseline has no value
When the baseline repository has no value, we can not normalize the
values relative to it. This is a problem, as all other ratios will
likely be around 1 and the raw value can have a wildly different value.

This patch now normalizes to the first repository with a value if the
baseline repository has no value.