[frontend] Fix detail graph loading
After a dependency update, the arrow function "findFirstSuccessful"
apparently bound to a different "this" - one *before* the component was
properly initialized. This was bad, as it did not receive any reactive
getters and was reading outdated values.

This PR converts it to a normal function, which properly captures the
this context.