[frontend] Hide shaded background on significance chips
The "to" property allows Vuetify to perform "active" tracking: When the
URL the user is on matches the url in the "to" property, the element
will become active and highlighted. Sadly, this is quite annoying as we
do not want our significance chips to suddenly light up on the
comparison page.

Vuetify does not offer a way to disable that as of now, so we just
manually force the background to be transparent.

Upstream bug:
https://github.com/vuetifyjs/vuetify/issues/11149