[frontend] Fix clicking on rows in run detail table
Clicking on a row should take you to the detail page with the given row
selected as reference and centered in the graph. This was broken by some
dependency update, as the setters apparently weren't correctly invoked
after it. This patch fixes that by explicitly using vxm instead of
"this" to invoke these explicit setters.