[frontend] Disable some no-console warnings
While there is probably a good reason for the warning to exist, we want
to at least log some errors before the rest of the application has
bootstrapped enough to display visuals (e.g. a snackbar).
The errors are non-critical but might provide us with some guidance when
debugging or give the user some explanation for unexpected program
behaviour.

Additionally, there are a few info logs which help explain the restore
logic. They are quite unobtrusive but might provide some useful insight.