[frontend] Add configuration for location independent image
A location independent image uses relative URLs in the bundle and
additionally tries to figure out at runtime where it is hosted to
configure the router and axios accordingly.

It does that by looking at window.location and chopping off any known
paths:
 - `https://example.com/velcom/home` would result in `/velcom` as the
   known route (`/home`) is cut off.
 - `https://example.com/velcom/compare/foo/to/bar` would result in
   `/velcom` as the known route (`/compare`) along with its parameters
   (`foo/to/bar`) is cut off.

Additionally `/api` is appended to compute the URL for backend requests.

This, in conjunction with a properly configured backend, allows the
VelCom page to be deployed at *any subpath* without needing explicit
configuration. Such a backend will be introduced in the following
commits in the built docker image.