[frontend] Only render run detail component once
Vue (intelligently) re-uses components when using v-if. In this case,
this lead to vue re-using the run-detail row for the commit component
once the commit was fetched.
This, in turn, triggered a re-render of the run-detail page including
the relatively expensive measurement table inside.

Giving either the commit view (or tar) or the run-detail row a unique
key prevents this re-use and only renders the table once. This patch
adds key to all rows for good measure, but the other components are
relatively inexpensive.